Push all the body parsing code into body.ts
